[python] refactor ray merge_into module
