+++ paimon-python/pypaimon/ray/__init__.py
@@ -16,6 +16,6 @@
 # under the License.
 
 from pypaimon.ray.ray_paimon import read_paimon, write_paimon
-from pypaimon.ray.merge_into import merge_paimon
+from pypaimon.ray.data_evolution_merge_into import merge_into
 
-__all__ = ["read_paimon", "write_paimon", "merge_paimon"]
+__all__ = ["read_paimon", "write_paimon", "merge_into"]
